Move Common code to lib library
