+++ lib/Common/sensors_state.h
