+++ platformio.ini
@@ -126,17 +126,15 @@ build_flags =
 	-O0
 build_unflags =
 	-Os
-lib_extra_dirs =
-	src_common
 
 [env:test]
-platform = native@1.2.1
-build_flags =	-std=gnu++11  -include test/mock.h
+platform = native
+build_flags =
+	-std=gnu++17
+	-g -O0
 lib_deps =
 	ArduinoFake
-test_build_src = true
-lib_extra_dirs =
-	src_common
+	powerbroker2/SerialTransfer
 
 [env:webserver]
 platform = https://github.com/platformio/platform-espressif32.git
@@ -169,12 +167,6 @@ extra_scripts =
 	pre:scripts/custom-src-dir.py
 	webserver/replace_fs.py
 
-lib_extra_dirs =
-	src_common
-
-build_unflags =
-	-std=gnu++11
-
 [env:uart-esp]
 framework = arduino
 platform = https://github.com/platformio/platform-espressif32.git
@@ -191,9 +183,6 @@ lib_deps =
 extra_scripts =
 	pre:scripts/custom-src-dir.py
 	webserver/replace_fs.py
-
-lib_extra_dirs =
-	src_common
 
 build_flags =
 	-std=gnu++17
@@ -217,9 +206,6 @@ lib_deps =
 
 extra_scripts =
 	pre:scripts/custom-src-dir.py
-
-lib_extra_dirs =
-	src_common
 
 build_flags =
 	-DPIO_FRAMEWORK_ARDUINO_ENABLE_CDC
